Introduce Connect_Retry_On_Refused procedure
The procedure tries to connect to a socket given by path. If the socket
responds with connection refused, sleep one second and retry. This might
happen if the manager created the socket but is not yet ready to accept
connections. Raises custom Connection_Refused exception if socket
refuses connection after Count tries.